Mudança na organização das pastas e do nome da interface ClassroomServices
