+++ src/main/java/com/example/qsproject/qsproject/DataImporter.java
@@ -13,7 +13,8 @@ public class DataImporter {
 
     @PostConstruct
     public void runImport() {
-        String csvFilePath = "./FilesForUs/filldb.csv";
+       // String csvFilePath = "./FilesForUs/filldb.csv";
+        String csvFilePath = "FilesForUs\\filldb.csv";
         classroomServices.importCSV(csvFilePath);
     }
 }
