+++ src/main/java/com/example/qsproject/qsproject/QsprojectApplication.java
@@ -6,7 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class QsprojectApplication {
 
-	public static void main(String[] args) {
+	public static void main(String[] args)
+	{
 		SpringApplication.run(QsprojectApplication.class, args);
 	}
 }
